Move "sl" fixture to conftest.py
